Fixed Break to IntroScene Caused by Change to MoveToMenu Script, Also Fixed Loading Screen Scaling
Player was not able to click to continue in Intro Scene because the loadSceneController GameObject has been using the MoveToMenu script which was just recently changed to have 6 new member variables which were left uninitialized in IntroScene. Also, the black background for loading screens was set to just display in the middle of the screen at the image's original resolution (751x422) which on most monitors would not fill the screen, which was the intended behavior. To solve this I set the background to stretch to fill the screen via a built-in Unity setting.